final push for submission

+++ schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE products
 	item_id int NOT NULL AUTO_INCREMENT,
 	product_name varchar(255) NOT NULL,
 	department_name varchar(255) NOT NULL,
-	price int NOT NULL,
+	price decimal(15,2) NOT NULL,
 	stock_quantity int NOT NULL,
 	PRIMARY KEY (item_id)
 );
